simplify eslint
