+++ packages/insomnia/src/common/project.ts
@@ -2,14 +2,10 @@ import {
   type ApiSpec,
   database,
   type GitRepository,
-  type GrpcRequest,
   type MockServer,
   models,
   type Project,
-  type Request,
   services,
-  type SocketIORequest,
-  type WebSocketRequest,
   type Workspace,
   type WorkspaceMeta,
   type WorkspaceScope,
@@ -91,109 +87,6 @@ const lockGenerator = () => {
 // otherwise they may interfere with each other, which may cause duplicate projects or other inconsistencies.
 // TODO: move all project operations to this file to ensure they are properly wrapped with locks
 export const projectLock = lockGenerator();
-
-type TrackableRecentRequest = Request | WebSocketRequest | GrpcRequest | SocketIORequest;
-
-export interface RecentProjectRequest {
-  workspaceId: string;
-  request: TrackableRecentRequest;
-}
-
-interface CachedProjectRecentRequest {
-  requestId: string;
-  workspaceId: string;
-}
-
-interface CachedProjectRecentRequestsPayload {
-  recentRequests: CachedProjectRecentRequest[];
-}
-
-const MAX_RECENT_PROJECT_REQUESTS = 5;
-const RECENT_PROJECT_REQUESTS_STORAGE_KEY_PREFIX = 'recent-project-requests';
-
-const getRecentProjectRequestsStorageKey = (projectId: string) =>
-  `${RECENT_PROJECT_REQUESTS_STORAGE_KEY_PREFIX}:${projectId}`;
-
-const removeCachedProjectRecentRequests = (projectId: string) => {
-  window.localStorage.removeItem(getRecentProjectRequestsStorageKey(projectId));
-};
-
-export const getCachedProjectRecentRequests = (projectId: string): CachedProjectRecentRequest[] => {
-  try {
-    const storedRecentRequests = window.localStorage.getItem(getRecentProjectRequestsStorageKey(projectId));
-
-    if (!storedRecentRequests) {
-      return [];
-    }
-
-    const payload = JSON.parse(storedRecentRequests) as Partial<CachedProjectRecentRequestsPayload>;
-
-    if (!payload || typeof payload !== 'object' || !Array.isArray(payload.recentRequests)) {
-      removeCachedProjectRecentRequests(projectId);
-      return [];
-    }
-
-    return payload.recentRequests.slice(0, MAX_RECENT_PROJECT_REQUESTS);
-  } catch {
-    removeCachedProjectRecentRequests(projectId);
-    return [];
-  }
-};
-
-export const recordProjectRecentRequest = ({
-  projectId,
-  requestId,
-  workspaceId,
-}: {
-  projectId: string;
-  requestId: string;
-  workspaceId: string;
-}) => {
-  if (!projectId || !requestId || !workspaceId) {
-    return;
-  }
-
-  const existingRecentRequests = getCachedProjectRecentRequests(projectId);
-  const payload: CachedProjectRecentRequestsPayload = {
-    recentRequests: [
-      { requestId, workspaceId },
-      ...existingRecentRequests.filter(storedRequest => storedRequest.requestId !== requestId),
-    ].slice(0, MAX_RECENT_PROJECT_REQUESTS),
-  };
-
-  window.localStorage.setItem(getRecentProjectRequestsStorageKey(projectId), JSON.stringify(payload));
-};
-
-export const getProjectRecentRequests = async (projectId: string) => {
-  const cachedRecentRequests = getCachedProjectRecentRequests(projectId);
-
-  if (!projectId || cachedRecentRequests.length === 0) {
-    return [];
-  }
-
-  const recentRequests = (
-    await Promise.all(
-      cachedRecentRequests.map(async ({ requestId, workspaceId }): Promise<RecentProjectRequest | null> => {
-        try {
-          const request = (await services.helpers.getRequestById(requestId)) as TrackableRecentRequest | null;
-
-          if (!request) {
-            return null;
-          }
-
-          return {
-            workspaceId,
-            request,
-          };
-        } catch {
-          return null;
-        }
-      }),
-    )
-  ).filter(isNotNullOrUndefined);
-
-  return recentRequests;
-};
 
 export const checkSingleProjectSyncStatus = async (projectId: string) => {
   const projectWorkspaces = await services.workspace.findByParentId(projectId);
@@ -312,80 +205,8 @@ export async function getAllLocalFiles({ projectId }: { projectId: string }) {
   return files;
 }
 
-export const getAllRemoteBackendProjectsByProjectId = async ({
-  teamProjectId,
-  organizationId,
-}: {
-  teamProjectId: string;
-  organizationId: string;
-}) => {
-  return window.main.sync.remoteBackendProjects({ teamId: organizationId, teamProjectId });
-};
-
-export const getAllRemoteBackendProjectsOfOrg = async ({ organizationId }: { organizationId: string }) => {
-  return window.main.sync.remoteBackendProjectsOfTeam({ teamId: organizationId });
-};
-
 export const getUnsyncedRemoteWorkspaces = (remoteFiles: InsomniaFile[], workspaces: Workspace[]) =>
   remoteFiles.filter(remoteFile => !workspaces.find(w => w._id === remoteFile.id));
-
-export async function getAllRemoteFiles({ projectId, organizationId }: { projectId: string; organizationId: string }) {
-  try {
-    const project = await services.project.get(projectId);
-
-    const remoteId = project?.remoteId;
-    if (!remoteId) {
-      return [];
-    }
-
-    console.log(
-      '[getAllRemoteFiles] start fetching remote backend workspaces for project',
-      projectId,
-      `remoteId: ${remoteId}`,
-    );
-
-    const [allPulledBackendProjectsForRemoteId, allFetchedRemoteBackendProjectsForRemoteId] = await Promise.all([
-      window.main.sync.localBackendProjects().then(projects => projects.filter(p => p.id === remoteId)),
-      // Remote backend projects are fetched from the backend since they are not stored locally
-      window.main.sync.remoteBackendProjects({ teamId: organizationId, teamProjectId: remoteId }),
-    ]);
-    console.log(
-      `[getAllRemoteFiles] found allPulledBackendProjectsForRemoteId: ${allPulledBackendProjectsForRemoteId.length} and allFetchedRemoteBackendProjectsForRemoteId: ${allFetchedRemoteBackendProjectsForRemoteId.length} for remoteId: ${remoteId}`,
-    );
-    // Get all workspaces that are connected to backend projects and under the current project
-    const workspacesWithBackendProjects = await database.find<Workspace>(models.workspace.type, {
-      _id: {
-        $in: [...allPulledBackendProjectsForRemoteId, ...allFetchedRemoteBackendProjectsForRemoteId].map(
-          p => p.rootDocumentId,
-        ),
-      },
-      parentId: project._id,
-    });
-    console.log(`[getAllRemoteFiles] found workspacesWithBackendProjects: ${workspacesWithBackendProjects.length}`);
-    // Get the list of remote backend projects that we need to pull
-    const backendProjectsToPull = allFetchedRemoteBackendProjectsForRemoteId.filter(
-      p => !workspacesWithBackendProjects.find(w => w._id === p.rootDocumentId),
-    );
-    console.log(`[getAllRemoteFiles] get ${backendProjectsToPull.length} unsynced files`);
-    return backendProjectsToPull.map(backendProject => {
-      const file: InsomniaFile = {
-        id: backendProject.rootDocumentId,
-        name: backendProject.name,
-        scope: 'unsynced',
-        label: 'Unsynced',
-        remoteId: backendProject.id,
-        created: 0,
-        lastModifiedTimestamp: 0,
-      };
-
-      return file;
-    });
-  } catch (e) {
-    console.warn('Failed to load backend projects', e);
-  }
-
-  return [];
-}
 
 /**
  * Get all projects for an organization with their associated git repositories
