+++ packages/insomnia/src/routes/organization.$organizationId.project.$projectId.tsx
@@ -9,12 +9,7 @@ import * as reactUse from 'react-use';
 import { logout } from '~/account/session';
 import { Icon } from '~/basic-components/icon';
 import { DEFAULT_SIDEBAR_SIZE } from '~/common/constants';
-import {
-  checkAllProjectSyncStatus,
-  getAllLocalFiles,
-  getAllRemoteFiles,
-  getProjectsWithGitRepositories,
-} from '~/common/project';
+import { checkAllProjectSyncStatus, getAllLocalFiles, getProjectsWithGitRepositories } from '~/common/project';
 import { useStorageRulesLoaderFetcher } from '~/routes/organization.$organizationId.storage-rules';
 import { ProjectModal } from '~/ui/components/modals/project-modal';
 import { ScratchPadTutorialPanel } from '~/ui/components/panes/scratchpad-tutorial-pane';
@@ -29,6 +24,7 @@ import { GitFileIssuesProvider, useProjectGitFileIssues } from '~/ui/hooks/use-g
 import { useLoaderDeferData } from '~/ui/hooks/use-loader-defer-data';
 import { useOrganizationPermissions } from '~/ui/hooks/use-organization-features';
 import { DEFAULT_STORAGE_RULES } from '~/ui/organization-utils';
+import { getAllRemoteFiles } from '~/ui/utils/remote-projects';
 import { invariant } from '~/utils/invariant';
 
 import type { Route } from './+types/organization.$organizationId.project.$projectId';
