+++ packages/insomnia/src/ui/components/design-empty-state.tsx
@@ -2,8 +2,9 @@ import type { IconName } from '@fortawesome/fontawesome-svg-core';
 import React, { type FC } from 'react';
 import { Button, Heading, Menu, MenuItem, MenuTrigger, Popover } from 'react-aria-components';
 
-import { documentationLinks } from '../../common/documentation';
 import { selectFileOrFolder } from '~/ui/utils/select-file-or-folder';
+
+import { documentationLinks } from '../../common/documentation';
 import { blankSpec as emptySpec, petStoreSpec, todoSpec } from './example-openapi-specs';
 import { Icon } from './icon';
 import { showModal } from './modals';
