+++ packages/insomnia/src/ui/components/modals/workspace-settings-modal.tsx
@@ -170,7 +170,8 @@ export const WorkspaceSettingsModal = ({ workspace, gitFilePath, project, mockSe
                             className="w-full outline-hidden [grid-area:input] placeholder:italic focus:outline-hidden"
                           />
                           <span className="pointer-events-none truncate opacity-0 [grid-area:input]">
-                            {safeToUseInsomniaFileName(fileNameValue) || (workspace.name ? safeToUseInsomniaFileName(workspace.name) : 'name')}
+                            {safeToUseInsomniaFileName(fileNameValue) ||
+                              (workspace.name ? safeToUseInsomniaFileName(workspace.name) : 'name')}
                           </span>
                           <span className="text-(--hl) [grid-area:extension]">.yaml</span>
                         </div>
