+++ packages/insomnia/src/ui/components/panes/grpc-request-pane.tsx
@@ -5,11 +5,11 @@ import { Tab, TabList, TabPanel, Tabs } from 'react-aria-components';
 import { useParams } from 'react-router';
 import * as reactUse from 'react-use';
 
-import { recordProjectRecentRequest } from '~/ui/utils/recent-project-requests';
 import { useRootLoaderData } from '~/root';
 import { AnalyticsEvent } from '~/ui/analytics';
 import { CodeEditor, type CodeEditorHandle } from '~/ui/components/.client/codemirror/code-editor';
 import { OneLineEditor } from '~/ui/components/.client/codemirror/one-line-editor';
+import { recordProjectRecentRequest } from '~/ui/utils/recent-project-requests';
 
 import { getCommonHeaderNames, getCommonHeaderValues } from '../../../common/common-headers';
 import { database as db } from '../../../common/database';
